calculate peptide characteristics
