+++ api/src/org/labkey/api/protein/ProteinService.java
@@ -27,7 +27,6 @@ import org.xml.sax.SAXException;
 
 import javax.xml.parsers.ParserConfigurationException;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
@@ -75,7 +74,7 @@ public interface ProteinService
     List<QueryViewProvider<PeptideSearchForm>> getPeptideSearchViews();
 
     /** @param aaRowWidth the number of amino acids to display in a single row */
-    WebPartView<?> getProteinCoverageView(int seqId, List<PeptideCharacter> peptideCharacters, int aaRowWidth, boolean showEntireFragmentInCoverage, @Nullable String accessionForFeatures);
+    WebPartView<?> getProteinCoverageView(int seqId, List<PeptideCharacteristic> peptideCharacteristics, int aaRowWidth, boolean showEntireFragmentInCoverage, @Nullable String accessionForFeatures);
 
     /** @return a web part with all of the annotations and identifiers we know for a given protein */
     WebPartView<?> getAnnotationsView(int seqId, Map<String, Collection<HtmlString>> extraAnnotations);
